requirements.txt
bcrypt==3.2.0  
Click==7.0
cryptography==35.0.0
fastapi==0.70.0
gunicorn==20.1.0
h11==0.9.0
passlib==1.7.4
pipenv==11.9.0
pydantic==1.8.2
python-jose==3.3.0
python-multipart==0.0.5
starlette==0.16.0
typing-extensions==3.10.0.2
uvicorn==0.11.3
uvloop==0.14.0
virtualenv==20.0.17
virtualenv-clone==0.3.0